light: factor out LoggenExecutor and add Docker support
