+++ tests/light/src/axosyslog_light/fixtures.py
@@ -36,6 +36,9 @@ import pytest
 from axosyslog_light.common.file import copy_file
 from axosyslog_light.common.pytest_operations import calculate_testcase_name
 from axosyslog_light.helpers.loggen.loggen import Loggen
+from axosyslog_light.helpers.loggen.loggen_docker_executor import LoggenDockerExecutor
+from axosyslog_light.helpers.loggen.loggen_executor import LoggenExecutor
+from axosyslog_light.helpers.loggen.loggen_local_executor import LoggenLocalExecutor
 from axosyslog_light.message_builder.bsd_format import BSDFormat
 from axosyslog_light.message_builder.log_message import LogMessage
 from axosyslog_light.syslog_ng.syslog_ng import SyslogNg
@@ -266,6 +269,15 @@ def setup(request):
     copy_file(testcase_parameters.get_testcase_file(), Path.cwd())
     light_extra_files(Path.cwd())
     request.addfinalizer(lambda: logger.info("Report file path\n{}\n".format(calculate_report_file_path(Path.cwd()))))
+
+    if request.config.getoption("--runner") == "docker":
+        LoggenExecutor.set_default_executor(
+            LoggenDockerExecutor(request.config.getoption("--docker-image")),
+        )
+    elif request.config.getoption("--runner") == "local":
+        LoggenExecutor.set_default_executor(
+            LoggenLocalExecutor(Path(request.config.getoption("--installdir"), "bin", "loggen")),
+        )
 
 
 class PortAllocator():
